refactor to address comments
