+++ isaaclab_arena/agentic_environment_generation/spec_validation.py
@@ -14,7 +14,7 @@ from pydantic import ValidationError
 from isaaclab_arena.environment_spec.arena_env_graph_spec import ArenaEnvGraphSpec
 
 if TYPE_CHECKING:
-    from isaaclab_arena.agentic_environment_generation.environment_generation_agent import (
+    from isaaclab_arena.agentic_environment_generation.catalogues import (
         AssetCatalogue,
         RelationCatalogue,
         TaskCatalogue,
@@ -49,18 +49,23 @@ def collect_agent_ready_validation_trace(
 ) -> list[str]:
     """Return spec violations against the exact catalogues exposed to the agent."""
     traces: list[str] = []
+
+    # Check every registry_name against the asset catalogue the agent was given.
     asset_names = {
         "EMBODIMENTS": {entry["name"] for entry in asset_catalog.embodiments},
         "BACKGROUNDS": {entry["name"] for entry in asset_catalog.backgrounds},
         "OBJECTS": {entry["name"] for entry in asset_catalog.objects},
     }
+    # Embodiment and background are required singles; flag any name outside the catalogue.
     if spec.embodiment.registry_name not in asset_names["EMBODIMENTS"]:
         traces.append(f"Embodiment registry_name {spec.embodiment.registry_name!r} is not in the EMBODIMENTS catalog")
     if spec.background.registry_name not in asset_names["BACKGROUNDS"]:
         traces.append(f"Background registry_name {spec.background.registry_name!r} is not in the BACKGROUNDS catalog")
+    # Each movable object must resolve to a known OBJECTS entry.
     for obj in spec.objects:
         if obj.registry_name not in asset_names["OBJECTS"]:
             traces.append(f"Object {obj.id!r} registry_name {obj.registry_name!r} is not in the OBJECTS catalog")
+    # Object-set members are registry names too; validate each one the same way.
     for object_set in spec.object_sets or []:
         for member in object_set.members:
             if member not in asset_names["OBJECTS"]:
@@ -68,12 +73,15 @@ def collect_agent_ready_validation_trace(
                     f"Object set {object_set.id!r} member registry_name {member!r} is not in the OBJECTS catalog"
                 )
 
+    # Check each subtask kind and its params against the task catalogue.
     task_entries = {entry.name: entry for entry in task_catalog.tasks}
     for task in spec.task.subtasks:
         entry = task_entries.get(task.kind)
+        # Unknown task kinds cannot be repaired via params; skip further checks.
         if entry is None:
             traces.append(f"Task {task.kind!r} is not in the TASKS catalog")
             continue
+        # Required params must be present; anything outside required|optional is rejected.
         for required_param in entry.required_params:
             if required_param not in task.params:
                 traces.append(f"Task {task.kind!r} is missing required param {required_param!r}")
@@ -84,9 +92,11 @@ def collect_agent_ready_validation_trace(
                 f"supported params are {sorted(supported_params)!r}"
             )
 
+    # Check each relation kind and its params against the relation catalogue.
     relation_entries = {entry.name: entry for entry in relation_catalog.relations}
     for relation in spec.relations:
         entry = relation_entries.get(relation.kind)
+        # Same pattern as tasks: unknown kinds first, then required/unsupported params.
         if entry is None:
             traces.append(f"Relation {relation.kind!r} is not in the RELATIONS catalog")
             continue
