+++ isaaclab_arena/tests/test_environment_generation_agent.py
@@ -11,13 +11,13 @@ from unittest.mock import patch
 
 import pytest
 
-from isaaclab_arena.agentic_environment_generation.environment_generation_agent import (
+from isaaclab_arena.agentic_environment_generation.catalogues import (
     AssetCatalogue,
-    EnvironmentGenerationAgent,
     build_asset_catalogue,
     build_relation_catalogue,
     build_task_catalogue,
 )
+from isaaclab_arena.agentic_environment_generation.environment_generation_agent import EnvironmentGenerationAgent
 from isaaclab_arena.agentic_environment_generation.simready_asset_search import (
     SimReadyCandidateCatalogue,
     SimReadyObjectCandidate,
